Updated bash tool `update_dump_describes`
It is now grouping the commands that change directories instead of
using `pushd` and `popd`.

It's probably less performant as it spawns a subshell, but I've seen
`pushd` and `popd` fail in the older Bourne shell which npm spawns when
running a script.

Also, there's a clunkiness to those two built-ins that I don't like.