delegation.js: Simplified!
Removed `delegate` function. It was confusing how it was different
from `create` since it replicated much of the functionality.

Originally, my intention was to "opt-in" to the ability of object
"methods" to call "super" only by calling `delegate` (`create` would
behave more or less the same as Object.create).

But thinking about it more, it only confused things.  If someone
wants to "opt-out", just use Object.create as usual to not get the
"super" ability.